JavaScript exception mapping fixes
