+++ js/src/Ice/IPEndpointI.js
@@ -3,7 +3,7 @@
 //
 
 import { Address } from "./Address.js";
-import { EndpointParseException } from "./LocalException.js";
+import { EndpointParseException } from "./LocalExceptions.js";
 import { HashUtil } from "./HashUtil.js";
 import { StringUtil } from "./StringUtil.js";
 import { EndpointI } from "./EndpointI.js";
