+++ js/src/Ice/InstanceExtensions.js
@@ -17,8 +17,8 @@ import { RouterManager } from "./RouterManager.js";
 import { Timer } from "./Timer.js";
 import { TraceLevels } from "./TraceLevels.js";
 import { ValueFactoryManager } from "./ValueFactoryManager.js";
-import { LocalException } from "./Exception.js";
-import { CommunicatorDestroyedException, InitializationException } from "./LocalException.js";
+import { LocalException } from "./LocalException.js";
+import { CommunicatorDestroyedException, InitializationException } from "./LocalExceptions.js";
 import { getProcessLogger } from "./ProcessLogger.js";
 import { ToStringMode } from "./ToStringMode.js";
 import { ProtocolInstance } from "./ProtocolInstance.js";
