+++ js/src/Ice/LocatorInfo.js
@@ -9,8 +9,9 @@ import { Protocol } from "./Protocol.js";
 import { EndpointSelectionType } from "./EndpointSelectionType.js";
 import { Promise } from "./Promise.js";
 import { identityToString } from "./IdentityUtil.js";
-import { LocalException, UserException } from "./Exception.js";
-import { NotRegisteredException } from "./LocalException.js";
+import { LocalException } from "./LocalException.js";
+import { UserException } from "./UserException.js";
+import { NotRegisteredException } from "./LocalExceptions.js";
 import { Debug } from "./Debug.js";
 
 export class LocatorInfo {
