+++ js/src/Ice/Operation.js
@@ -5,7 +5,7 @@
 import { Ice as Ice_OperationMode } from "./OperationMode.js";
 const { OperationMode } = Ice_OperationMode;
 import { FormatType } from "./FormatType.js";
-import { UnknownException, MarshalException, OperationNotExistException } from "./LocalException.js";
+import { UnknownException, MarshalException, OperationNotExistException } from "./LocalExceptions.js";
 import { ObjectPrx } from "./ObjectPrx.js";
 import { TypeRegistry } from "./TypeRegistry.js";
