+++ js/src/Ice/Reference.js
@@ -15,7 +15,7 @@ import { OpaqueEndpointI } from "./OpaqueEndpoint.js";
 import { ReferenceMode } from "./ReferenceMode.js";
 import { StringUtil } from "./StringUtil.js";
 import { Encoding_1_0, Protocol_1_0, encodingVersionToString, protocolVersionToString } from "./Protocol.js";
-import { FixedProxyException, NoEndpointException } from "./LocalException.js";
+import { FixedProxyException, NoEndpointException } from "./LocalExceptions.js";
 import { Promise } from "./Promise.js";
 import { Debug } from "./Debug.js";
 import { ConnectRequestHandler } from "./ConnectRequestHandler.js";
