+++ js/src/Ice/WSTransceiver.js
@@ -2,7 +2,7 @@
 // Copyright (c) ZeroC, Inc. All rights reserved.
 //
 
-import { ConnectFailedException, ConnectionLostException, SocketException } from "./LocalException.js";
+import { ConnectFailedException, ConnectionLostException, SocketException } from "./LocalExceptions.js";
 import { WSConnectionInfo, TCPConnectionInfo } from "./Connection.js";
 import { ConnectionInfo as SSLConnectionInfo } from "./SSL/ConnectionInfo.js";
 import { SocketOperation } from "./SocketOperation.js";
