+++ js/src/Ice/index.d.ts
@@ -30,6 +30,7 @@
 /// <reference path="./Context.d.ts" />
 /// <reference path="./Logger.d.ts" />
 /// <reference path="./LocalException.d.ts" />
+/// <reference path="./LocalExceptions.d.ts" />
 /// <reference path="./Protocol.d.ts" />
 /// <reference path="./IdentityUtil.d.ts" />
 /// <reference path="./Current.d.ts" />
@@ -48,4 +49,5 @@
 /// <reference path="./AsyncResult.d.ts" />
 /// <reference path="./ServantLocator.d.ts" />
 /// <reference path="./UnknownSlicedValue.d.ts" />
+/// <reference path="./UserException.d.ts" />
 /// <reference path="./AsyncResultBase.d.ts" />
